Improved performance of AsRaw (Value) and AsValue (Raw).
Previous implementation caused lots of garbage collections when dealing with large streams. Both nodes reuse the output streams now instead of creating new ones.
They both disable the AutoFlush behaviour of the plugin host and flush the output manually in order to mark it as changed, because only the stream data changes not the stream classes.