Changed I/O pin namings
Changed Shape to Polygon, as this is more sprecific

Changed Selected to Tagged, as Select (PolygonEditor) would be easily
confused with Select (Value), which has completely different
functionality

Changed Shape Vertex Count to Vertex Bin Size, as the number of Vertices
per Polygon (input spreadable)

Changed Shape Z-Index to Z-Index, as it is clear we are dealing with
shapes (sorry, polygons). Same with all other pins carrying properties
of shapes.