Included SnapBounds
note that SnapBounds can only handle one tagged polygon at a time (a
polygon is tagged when all it's vertices are tagged). If two polygons
are tagged, it will take the first one in the spread. Multiple selection
is turned off for now in Shape Mode. Also, without Force Shape Mode,
only one polygon at a time can be in Shape Mode, which is just for
convenience.