ApplyTransform (Mouse) bugfix
Using getSlice to only get the first transform, in case there are
multiple slices coming in.

anything else does not really make sense (Most of the time, it would
just use the first transform, anyway. Only when multiple MouseEvents
come in, the second, third... transform would be used, which is very
confusing.)