[VVVV] Fixed the SetSlice plugin (issue #2656)
Turns out the SetSlice plugin was broken from the beginning. The FOutput.AssignFrom(FSpread) call at the beginning assigned the input spreads to the output, but it did not make a deep copy of those spreads. Later writes to the output did therefor overwrite the input.
The issue didn't surface before because in older versions of vvvv the transform pins always reported that they're changed so the input of the SetSlice plugin had fresh data every new frame.